Tidy Data Project Alpha Update
All of the required files are completed to the best of my own ability.
The program now returns what I believe to be a tidy data set. The
Readme explains how this is done. The Codebook describes all of the
variables. Finally, a text file output is uploaded to show what the
program can do.